tools: Move objparser from demo/ to tools/
This seems like a more appropriate location; that way demo/ depends on
tools/ and not the other way around.

Also add BUILD_TOOLS to CMake to build meshencoder.
