+++ CMakeLists.txt
@@ -2,6 +2,7 @@ project(meshoptimizer)
 cmake_minimum_required(VERSION 3.0)
 
 option(BUILD_DEMO "Build demo" OFF)
+option(BUILD_TOOLS "Build tools" OFF)
 
 set(SOURCES
     src/meshoptimizer.h
@@ -29,6 +30,11 @@ else()
 endif()
 
 if(BUILD_DEMO)
-    add_executable(demo demo/main.cpp demo/miniz.cpp demo/objparser.cpp)
+    add_executable(demo demo/main.cpp demo/miniz.cpp tools/objparser.cpp)
     target_link_libraries(demo meshoptimizer)
+endif()
+
+if(BUILD_TOOLS)
+    add_executable(meshencoder tools/meshencoder.cpp tools/objparser.cpp)
+    target_link_libraries(meshencoder meshoptimizer)
 endif()
