+++ Makefile
@@ -9,10 +9,13 @@ BUILD=build/$(config)
 LIBRARY_SOURCES=$(wildcard src/*.cpp)
 LIBRARY_OBJECTS=$(LIBRARY_SOURCES:%=$(BUILD)/%.o)
 
-DEMO_SOURCES=$(wildcard demo/*.c demo/*.cpp)
+DEMO_SOURCES=$(wildcard demo/*.c demo/*.cpp tools/objparser.cpp)
 DEMO_OBJECTS=$(DEMO_SOURCES:%=$(BUILD)/%.o)
 
-OBJECTS=$(LIBRARY_OBJECTS) $(DEMO_OBJECTS)
+ENCODER_SOURCES=tools/meshencoder.cpp tools/objparser.cpp
+ENCODER_OBJECTS=$(ENCODER_SOURCES:%=$(BUILD)/%.o)
+
+OBJECTS=$(LIBRARY_OBJECTS) $(DEMO_OBJECTS) $(ENCODER_OBJECTS)
 
 LIBRARY=$(BUILD)/libmeshoptimizer.a
 EXECUTABLE=$(BUILD)/meshoptimizer
@@ -50,11 +53,11 @@ dev: $(EXECUTABLE)
 format:
 	clang-format -i $(SOURCES)
 
-meshencoder: $(LIBRARY) $(BUILD)/demo/objparser.cpp.o tools/meshencoder.cpp
-	$(CXX) tools/meshencoder.cpp $(LIBRARY) $(BUILD)/demo/objparser.cpp.o $(CXXFLAGS) $(LDFLAGS) -o meshencoder
+meshencoder: $(ENCODER_OBJECTS) $(LIBRARY)
+	$(CXX) $^ $(LDFLAGS) -o $@
 
 $(EXECUTABLE): $(DEMO_OBJECTS) $(LIBRARY)
-	$(CXX) $(DEMO_OBJECTS) $(LIBRARY) $(LDFLAGS) -o $@
+	$(CXX) $^ $(LDFLAGS) -o $@
 
 $(LIBRARY): $(LIBRARY_OBJECTS)
 	ar rcs $@ $^
