+++ tools/objparser.cpp
