+++ tools/vcachetuner.cpp
@@ -1,5 +1,5 @@
-#include "../demo/objparser.h"
 #include "../src/meshoptimizer.h"
+#include "objparser.h"
 
 #include <algorithm>
 #include <functional>
